Refactor modifiers
