+++ src/main/node/EventListener.java
@@ -1,5 +1,7 @@
 package main.node;
 
+import main.node.Model.Event;
+
 public interface EventListener {
     void update(Event event);
 }
