+++ src/main/node/Main.java
@@ -1,5 +1,7 @@
 package main.node;
 
+import main.node.Model.Config;
+
 import java.net.InetAddress;
 import java.net.UnknownHostException;
 import java.util.Scanner;
