+++ src/main/node/Model/MessageType.java
@@ -1,4 +1,4 @@
-package main.node;
+package main.node.Model;
 
 public enum MessageType {
     BROADCAST, REGISTER, ID, NODE_LIST, OPEN_REQUEST, DATA,
