+++ src/main/node/Node.java
@@ -1,5 +1,10 @@
 package main.node;
 
+import main.node.Listeners.TcpListener;
+import main.node.Listeners.UdpListener;
+import main.node.Model.Event;
+import main.node.Model.Message;
+
 import java.io.IOException;
 import java.net.*;
